Add tests for run function generation; Extract tests data to a separate object
